requirements.txt
numpy
pycocotools
absl-py
scikit-learn
matplotlib
pillow